binding/c: autogen datatype api
