+++ src/binding/fortran/use_mpi/create_f90_util.c
@@ -105,7 +105,7 @@ int MPIR_Create_unnamed_predefined(MPI_Datatype old, int combiner,
             MPI_Datatype new_basic = MPI_DATATYPE_NULL;
             /* we used MPIR_Type_contiguous and then stomped it's contents
              * information, so make sure that the basic_type is usable by
-             * MPIR_Type_commit */
+             * MPIR_Type_commit_impl */
             MPIR_Datatype_get_basic_type(old, old_basic);
             MPIR_Datatype_get_basic_type(new_dtp->handle, new_basic);
             MPIR_Assert(new_basic == old_basic);
@@ -114,7 +114,7 @@ int MPIR_Create_unnamed_predefined(MPI_Datatype old, int combiner,
 
         /* the MPI Standard requires that these types are pre-committed
          * (MPI-2.2, sec 16.2.5, pg 492) */
-        mpi_errno = MPIR_Type_commit(&type->d);
+        mpi_errno = MPIR_Type_commit_impl(&type->d);
         MPIR_ERR_CHECK(mpi_errno);
     }
