+++ src/include/mpir_datatype.h
@@ -530,34 +530,39 @@ int MPIR_Type_create_indexed_block_impl(int count,
                                         int blocklength,
                                         const int array_of_displacements[],
                                         MPI_Datatype oldtype, MPI_Datatype * newtype);
+int MPIR_Type_create_hvector_impl(int count, int blocklength, MPI_Aint stride,
+                                  MPI_Datatype oldtype, MPI_Datatype * newtype);
 int MPIR_Type_create_hindexed_block_impl(int count, int blocklength,
                                          const MPI_Aint array_of_displacements[],
                                          MPI_Datatype oldtype, MPI_Datatype * newtype);
-int MPIR_Type_create_darray(int size, int rank, int ndims,
-                            const int array_of_gsizes[], const int array_of_distribs[],
-                            const int array_of_dargs[], const int array_of_psizes[],
-                            int order, MPI_Datatype oldtype, MPI_Datatype * newtype);
-int MPIR_Type_create_subarray(int ndims, const int array_of_sizes[],
-                              const int array_of_subsizes[], const int array_of_starts[],
-                              int order, MPI_Datatype oldtype, MPI_Datatype * newtype);
+int MPIR_Type_create_hindexed_impl(int count, const int array_of_blocklengths[],
+                                   const MPI_Aint array_of_displacements[],
+                                   MPI_Datatype oldtype, MPI_Datatype * newtype);
+int MPIR_Type_create_darray_impl(int size, int rank, int ndims,
+                                 const int array_of_gsizes[], const int array_of_distribs[],
+                                 const int array_of_dargs[], const int array_of_psizes[],
+                                 int order, MPI_Datatype oldtype, MPI_Datatype * newtype);
+int MPIR_Type_create_subarray_impl(int ndims, const int array_of_sizes[],
+                                   const int array_of_subsizes[], const int array_of_starts[],
+                                   int order, MPI_Datatype oldtype, MPI_Datatype * newtype);
 int MPIR_Type_contiguous_impl(int count, MPI_Datatype old_type, MPI_Datatype * new_type_p);
 int MPIR_Type_contiguous_x_impl(MPI_Count count, MPI_Datatype old_type, MPI_Datatype * new_type_p);
+int MPIR_Type_dup_impl(MPI_Datatype oldtype, MPI_Datatype * newtype);
+int MPIR_Type_create_resized_impl(MPI_Datatype oldtype, MPI_Aint lb, MPI_Aint extent,
+                                  MPI_Datatype * newtype);
 void MPIR_Type_get_extent_impl(MPI_Datatype datatype, MPI_Aint * lb, MPI_Aint * extent);
 void MPIR_Type_get_true_extent_impl(MPI_Datatype datatype, MPI_Aint * true_lb,
                                     MPI_Aint * true_extent);
 void MPIR_Type_get_envelope(MPI_Datatype datatype, int *num_integers, int *num_addresses,
                             int *num_datatypes, int *combiner);
-int MPIR_Type_hvector_impl(int count, int blocklen, MPI_Aint stride, MPI_Datatype old_type,
-                           MPI_Datatype * newtype_p);
 int MPIR_Type_indexed_impl(int count, const int blocklens[], const int indices[],
                            MPI_Datatype old_type, MPI_Datatype * newtype);
 void MPIR_Type_free_impl(MPI_Datatype * datatype);
 int MPIR_Type_vector_impl(int count, int blocklength, int stride, MPI_Datatype old_type,
                           MPI_Datatype * newtype_p);
-int MPIR_Type_struct_impl(int count, const int blocklens[], const MPI_Aint indices[],
-                          const MPI_Datatype old_types[], MPI_Datatype * newtype);
 void MPIR_Pack_size_impl(int incount, MPI_Datatype datatype, MPI_Aint * size);
 void MPIR_Type_lb_impl(MPI_Datatype datatype, MPI_Aint * displacement);
+int MPIR_Type_match_size_impl(int typeclass, int size, MPI_Datatype * datatype);
 
 /* Datatype functions */
 int MPIR_Type_dup(MPI_Datatype oldtype, MPI_Datatype * newtype);
@@ -573,13 +578,12 @@ int MPIR_Type_blockindexed(int count,
                            const void *displacement_array,
                            int dispinbytes, MPI_Datatype oldtype, MPI_Datatype * newtype);
 
-int MPIR_Type_commit(MPI_Datatype * type);
 int MPII_Type_zerolen(MPI_Datatype * newtype);
 int MPIR_Type_create_resized(MPI_Datatype oldtype, MPI_Aint lb, MPI_Aint extent,
                              MPI_Datatype * newtype);
-int MPIR_Type_get_contents(MPI_Datatype datatype, int max_integers, int max_addresses,
-                           int max_datatypes, int array_of_integers[],
-                           MPI_Aint array_of_addresses[], MPI_Datatype array_of_datatypes[]);
+int MPIR_Type_get_contents_impl(MPI_Datatype datatype, int max_integers, int max_addresses,
+                                int max_datatypes, int array_of_integers[],
+                                MPI_Aint array_of_addresses[], MPI_Datatype array_of_datatypes[]);
 int MPIR_Type_create_pairtype(MPI_Datatype datatype, MPIR_Datatype * new_dtp);
 
 /* debugging helper functions */
