+++ src/mpi/datatype/datatype_impl.c
@@ -38,316 +38,12 @@ void MPIR_Get_count_impl(const MPI_Status * status, MPI_Datatype datatype, MPI_A
     }
 }
 
-/* any non-MPI functions go here, especially non-static ones */
-
-/* any non-MPI functions go here, especially non-static ones */
-
-/* Internal helper routines.  If you want to get the number of elements from
- * within the MPI library, call MPIR_Get_elements_x_impl instead. */
-
-/* MPIR_Type_get_basic_type_elements()
- *
- * Arguments:
- * - bytes_p - input/output byte count
- * - count - maximum number of this type to subtract from the bytes; a count
- *           of -1 indicates use as many as we like
- * - datatype - input datatype
- *
- * Returns number of elements available given the two constraints of number of
- * bytes and count of types.  Also reduces the byte count by the amount taken
- * up by the types.
- *
- * Assumptions:
- * - the type passed to this function must be a basic *or* a pairtype
- *   (which aren't basic types)
- * - the count is not zero (otherwise we can't tell between a "no more
- *   complete types" case and a "zero count" case)
- *
- * As per section 4.9.3 of the MPI 1.1 specification, the two-part reduction
- * types are to be treated as structs of the constituent types.  So we have to
- * do something special to handle them correctly in here.
- *
- * As per section 3.12.5 get_count and get_elements report the same value for
- * basic datatypes; I'm currently interpreting this to *not* include these
- * reduction types, as they are considered structs.
- */
-static MPI_Count MPIR_Type_get_basic_type_elements(MPI_Count * bytes_p,
-                                                   MPI_Count count, MPI_Datatype datatype)
-{
-    MPI_Count elements, usable_bytes, used_bytes, type1_sz, type2_sz;
-
-    if (count == 0)
-        return 0;
-
-    /* determine the maximum number of bytes we should take from the
-     * byte count.
-     */
-    if (count < 0) {
-        usable_bytes = *bytes_p;
-    } else {
-        usable_bytes = MPL_MIN(*bytes_p, count * MPIR_Datatype_get_basic_size(datatype));
-    }
-
-    switch (datatype) {
-            /* we don't get valid fortran datatype handles in all cases... */
-#ifdef HAVE_FORTRAN_BINDING
-        case MPI_2REAL:
-            type1_sz = type2_sz = MPIR_Datatype_get_basic_size(MPI_REAL);
-            break;
-        case MPI_2DOUBLE_PRECISION:
-            type1_sz = type2_sz = MPIR_Datatype_get_basic_size(MPI_DOUBLE_PRECISION);
-            break;
-        case MPI_2INTEGER:
-            type1_sz = type2_sz = MPIR_Datatype_get_basic_size(MPI_INTEGER);
-            break;
-#endif
-        case MPI_2INT:
-            type1_sz = type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        case MPI_FLOAT_INT:
-            type1_sz = MPIR_Datatype_get_basic_size(MPI_FLOAT);
-            type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        case MPI_DOUBLE_INT:
-            type1_sz = MPIR_Datatype_get_basic_size(MPI_DOUBLE);
-            type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        case MPI_LONG_INT:
-            type1_sz = MPIR_Datatype_get_basic_size(MPI_LONG);
-            type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        case MPI_SHORT_INT:
-            type1_sz = MPIR_Datatype_get_basic_size(MPI_SHORT);
-            type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        case MPI_LONG_DOUBLE_INT:
-            type1_sz = MPIR_Datatype_get_basic_size(MPI_LONG_DOUBLE);
-            type2_sz = MPIR_Datatype_get_basic_size(MPI_INT);
-            break;
-        default:
-            /* all other types.  this is more complicated than
-             * necessary for handling these types, but it puts us in the
-             * same code path for all the basics, so we stick with it.
-             */
-            type1_sz = type2_sz = MPIR_Datatype_get_basic_size(datatype);
-            break;
-    }
-
-    /* determine the number of elements in the region */
-    elements = 2 * (usable_bytes / (type1_sz + type2_sz));
-    if (usable_bytes % (type1_sz + type2_sz) >= type1_sz)
-        elements++;
-
-    /* determine how many bytes we used up with those elements */
-    used_bytes = ((elements / 2) * (type1_sz + type2_sz));
-    if (elements % 2 == 1)
-        used_bytes += type1_sz;
-
-    *bytes_p -= used_bytes;
-
-    return elements;
-}
-
-
-/* MPIR_Type_get_elements
- *
- * Arguments:
- * - bytes_p - input/output byte count
- * - count - maximum number of this type to subtract from the bytes; a count
- *           of <0 indicates use as many as we like
- * - datatype - input datatype
- *
- * Returns number of elements available given the two constraints of number of
- * bytes and count of types.  Also reduces the byte count by the amount taken
- * up by the types.
- *
- * This is called from MPI_Get_elements() when it sees a type with multiple
- * element types (datatype_ptr->element_sz = -1).  This function calls itself too.
- */
-static MPI_Count MPIR_Type_get_elements(MPI_Count * bytes_p, MPI_Count count, MPI_Datatype datatype)
-{
-    MPIR_Datatype *datatype_ptr = NULL;
-
-    MPIR_Datatype_get_ptr(datatype, datatype_ptr);      /* invalid if builtin */
-
-    /* if we have gotten down to a type with only one element type,
-     * call MPIR_Type_get_basic_type_elements() and return.
-     */
-    if (HANDLE_IS_BUILTIN(datatype) ||
-        datatype == MPI_FLOAT_INT ||
-        datatype == MPI_DOUBLE_INT ||
-        datatype == MPI_LONG_INT || datatype == MPI_SHORT_INT || datatype == MPI_LONG_DOUBLE_INT) {
-        return MPIR_Type_get_basic_type_elements(bytes_p, count, datatype);
-    } else if (datatype_ptr->builtin_element_size >= 0) {
-        MPI_Datatype basic_type = MPI_DATATYPE_NULL;
-        MPIR_Datatype_get_basic_type(datatype_ptr->basic_type, basic_type);
-        return MPIR_Type_get_basic_type_elements(bytes_p,
-                                                 count * datatype_ptr->n_builtin_elements,
-                                                 basic_type);
-    } else {
-        /* we have bytes left and still don't have a single element size; must
-         * recurse.
-         */
-        int i, j, *ints;
-        MPI_Count typecount = 0, nr_elements = 0, last_nr_elements;
-        MPI_Aint *aints;
-        MPI_Datatype *types;
-
-        /* Establish locations of arrays */
-        MPIR_Type_access_contents(datatype_ptr->handle, &ints, &aints, &types);
-        if (!ints || !aints || !types)
-            return MPI_ERR_TYPE;
-
-        switch (datatype_ptr->contents->combiner) {
-            case MPI_COMBINER_NAMED:
-            case MPI_COMBINER_DUP:
-            case MPI_COMBINER_RESIZED:
-                return MPIR_Type_get_elements(bytes_p, count, *types);
-                break;
-            case MPI_COMBINER_CONTIGUOUS:
-            case MPI_COMBINER_VECTOR:
-            case MPI_COMBINER_HVECTOR_INTEGER:
-            case MPI_COMBINER_HVECTOR:
-            case MPI_COMBINER_SUBARRAY:
-                /* count is first in ints array */
-                return MPIR_Type_get_elements(bytes_p, count * (*ints), *types);
-                break;
-            case MPI_COMBINER_INDEXED_BLOCK:
-            case MPI_COMBINER_HINDEXED_BLOCK:
-                /* count is first in ints array, blocklength is second */
-                return MPIR_Type_get_elements(bytes_p, count * ints[0] * ints[1], *types);
-                break;
-            case MPI_COMBINER_INDEXED:
-            case MPI_COMBINER_HINDEXED_INTEGER:
-            case MPI_COMBINER_HINDEXED:
-                for (i = 0; i < (*ints); i++) {
-                    /* add up the blocklengths to get a max. # of the next type */
-                    typecount += ints[i + 1];
-                }
-                return MPIR_Type_get_elements(bytes_p, count * typecount, *types);
-                break;
-            case MPI_COMBINER_STRUCT_INTEGER:
-            case MPI_COMBINER_STRUCT:
-                /* In this case we can't simply multiply the count of the next
-                 * type by the count of the current type, because we need to
-                 * cycle through the types just as the struct would.  thus the
-                 * nested loops.
-                 *
-                 * We need to keep going until we get less elements than expected
-                 * or we run out of bytes.
-                 */
-
-
-                last_nr_elements = 1;   /* seed value */
-                for (j = 0; (count < 0 || j < count) && *bytes_p > 0 && last_nr_elements > 0; j++) {
-                    /* recurse on each type; bytes are reduced in calls */
-                    for (i = 0; i < (*ints); i++) {
-                        /* skip zero-count elements of the struct */
-                        if (ints[i + 1] == 0)
-                            continue;
-
-                        last_nr_elements = MPIR_Type_get_elements(bytes_p, ints[i + 1], types[i]);
-                        nr_elements += last_nr_elements;
-
-                        MPIR_Assert(last_nr_elements >= 0);
-
-                        if (last_nr_elements < ints[i + 1])
-                            break;
-                    }
-                }
-                return nr_elements;
-                break;
-            case MPI_COMBINER_DARRAY:
-            case MPI_COMBINER_F90_REAL:
-            case MPI_COMBINER_F90_COMPLEX:
-            case MPI_COMBINER_F90_INTEGER:
-            default:
-                /* --BEGIN ERROR HANDLING-- */
-                MPIR_Assert(0);
-                return -1;
-                break;
-                /* --END ERROR HANDLING-- */
-        }
-    }
-}
-
-/* MPIR_Get_elements_x_impl
- *
- * Arguments:
- * - byte_count - input/output byte count
- * - datatype - input datatype
- * - elements - Number of basic elements this byte_count would contain
- *
- * Returns number of elements available given the two constraints of number of
- * bytes and count of types.  Also reduces the byte count by the amount taken
- * up by the types.
- */
-int MPIR_Get_elements_x_impl(MPI_Count * byte_count, MPI_Datatype datatype, MPI_Count * elements)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Datatype *datatype_ptr = NULL;
-
-    if (!HANDLE_IS_BUILTIN(datatype)) {
-        MPIR_Datatype_get_ptr(datatype, datatype_ptr);
-    }
-
-    /* three cases:
-     * - nice, simple, single element type
-     * - derived type with a zero size
-     * - type with multiple element types (nastiest)
-     */
-    if (HANDLE_IS_BUILTIN(datatype) ||
-        (datatype_ptr->builtin_element_size != -1 && datatype_ptr->size > 0)) {
-        /* in both cases we do not limit the number of types that might
-         * be in bytes
-         */
-        if (!HANDLE_IS_BUILTIN(datatype)) {
-            MPI_Datatype basic_type = MPI_DATATYPE_NULL;
-            MPIR_Datatype_get_basic_type(datatype_ptr->basic_type, basic_type);
-            *elements = MPIR_Type_get_basic_type_elements(byte_count, -1, basic_type);
-        } else {
-            /* Behaves just like MPI_Get_Count in the predefined case */
-            MPI_Count size;
-            MPIR_Datatype_get_size_macro(datatype, size);
-            if ((*byte_count % size) != 0)
-                *elements = MPI_UNDEFINED;
-            else
-                *elements = MPIR_Type_get_basic_type_elements(byte_count, -1, datatype);
-        }
-        MPIR_Assert(*byte_count >= 0);
-    } else if (datatype_ptr->size == 0) {
-        if (*byte_count > 0) {
-            /* --BEGIN ERROR HANDLING-- */
-
-            /* datatype size of zero and count > 0 should never happen. */
-
-            (*elements) = MPI_UNDEFINED;
-            /* --END ERROR HANDLING-- */
-        } else {
-            /* This is ambiguous.  However, discussions on MPI Forum
-             * reached a consensus that this is the correct return
-             * value
-             */
-            (*elements) = 0;
-        }
-    } else {    /* derived type with weird element type or weird size */
-
-        MPIR_Assert(datatype_ptr->builtin_element_size == -1);
-
-        *elements = MPIR_Type_get_elements(byte_count, -1, datatype);
-    }
-
-    return mpi_errno;
-}
-
 void MPIR_Pack_size_impl(int incount, MPI_Datatype datatype, MPI_Aint * size)
 {
     MPI_Aint typesize;
     MPIR_Datatype_get_size_macro(datatype, typesize);
     *size = incount * typesize;
 }
-
-/* any non-MPI functions go here, especially non-static ones */
 
 int MPIR_Status_set_elements_x_impl(MPI_Status * status, MPI_Datatype datatype, MPI_Count count)
 {
@@ -367,7 +63,7 @@ int MPIR_Status_set_elements_x_impl(MPI_Status * status, MPI_Datatype datatype, 
     return mpi_errno;
 }
 
-int MPIR_Type_commit(MPI_Datatype * datatype_p)
+int MPIR_Type_commit_impl(MPI_Datatype * datatype_p)
 {
     int mpi_errno = MPI_SUCCESS;
     MPIR_Datatype *datatype_ptr;
@@ -388,28 +84,6 @@ int MPIR_Type_commit(MPI_Datatype * datatype_p)
 
     }
     return mpi_errno;
-}
-
-int MPIR_Type_commit_impl(MPI_Datatype * datatype)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (HANDLE_IS_BUILTIN(*datatype))
-        goto fn_exit;
-
-    /* pair types stored as real types are a special case */
-    if (*datatype == MPI_FLOAT_INT ||
-        *datatype == MPI_DOUBLE_INT ||
-        *datatype == MPI_LONG_INT || *datatype == MPI_SHORT_INT || *datatype == MPI_LONG_DOUBLE_INT)
-        goto fn_exit;
-
-    mpi_errno = MPIR_Type_commit(datatype);
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
 }
 
 int MPIR_Type_contiguous(int count, MPI_Datatype oldtype, MPI_Datatype * newtype)
@@ -821,6 +495,70 @@ static int MPIR_Type_cyclic(const int *array_of_gsizes,
     return MPI_SUCCESS;
 }
 
+int MPIR_Type_create_hvector_impl(int count, int blocklength, MPI_Aint stride,
+                                  MPI_Datatype oldtype, MPI_Datatype * newtype)
+{
+    int mpi_errno = MPI_SUCCESS;
+    MPI_Datatype new_handle;
+    MPIR_Datatype *new_dtp;
+    int ints[2];
+
+    mpi_errno = MPIR_Type_vector(count, blocklength, stride, 1, /* stride in bytes */
+                                 oldtype, &new_handle);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    ints[0] = count;
+    ints[1] = blocklength;
+    MPIR_Datatype_get_ptr(new_handle, new_dtp);
+    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_HVECTOR, 2,    /* ints (count, blocklength) */
+                                           1,   /* aints */
+                                           1,   /* types */
+                                           ints, &stride, &oldtype);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
+
+  fn_exit:
+    return mpi_errno;
+  fn_fail:
+    goto fn_exit;
+}
+
+int MPIR_Type_create_hindexed_impl(int count, const int array_of_blocklengths[],
+                                   const MPI_Aint array_of_displacements[],
+                                   MPI_Datatype oldtype, MPI_Datatype * newtype)
+{
+    int mpi_errno = MPI_SUCCESS;
+    MPI_Datatype new_handle;
+    MPIR_Datatype *new_dtp;
+    int *ints;
+    MPIR_CHKLMEM_DECL(1);
+
+    mpi_errno = MPIR_Type_indexed(count, array_of_blocklengths, array_of_displacements, 1,      /* displacements in bytes */
+                                  oldtype, &new_handle);
+    MPIR_ERR_CHECK(mpi_errno);
+    MPIR_CHKLMEM_MALLOC_ORJUMP(ints, int *, (count + 1) * sizeof(int), mpi_errno,
+                               "content description", MPL_MEM_BUFFER);
+    ints[0] = count;
+
+    for (int i = 0; i < count; i++) {
+        ints[i + 1] = array_of_blocklengths[i];
+    }
+    MPIR_Datatype_get_ptr(new_handle, new_dtp);
+    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_HINDEXED, count + 1,   /* ints (count, blocklengths) */
+                                           count,       /* aints (displacements) */
+                                           1,   /* types */
+                                           ints, array_of_displacements, &oldtype);
+    MPIR_ERR_CHECK(mpi_errno);
+    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
+
+  fn_exit:
+    MPIR_CHKLMEM_FREEALL();
+    return mpi_errno;
+  fn_fail:
+    goto fn_exit;
+}
+
 int MPIR_Type_create_hindexed_block_impl(int count, int blocklength,
                                          const MPI_Aint array_of_displacements[],
                                          MPI_Datatype oldtype, MPI_Datatype * newtype)
@@ -1010,10 +748,10 @@ int MPIR_Type_create_struct_impl(int count,
     goto fn_exit;
 }
 
-int MPIR_Type_create_darray(int size, int rank, int ndims,
-                            const int array_of_gsizes[], const int array_of_distribs[],
-                            const int array_of_dargs[], const int array_of_psizes[],
-                            int order, MPI_Datatype oldtype, MPI_Datatype * newtype)
+int MPIR_Type_create_darray_impl(int size, int rank, int ndims,
+                                 const int array_of_gsizes[], const int array_of_distribs[],
+                                 const int array_of_dargs[], const int array_of_psizes[],
+                                 int order, MPI_Datatype oldtype, MPI_Datatype * newtype)
 {
     int mpi_errno = MPI_SUCCESS;
     int i;
@@ -1235,9 +973,9 @@ int MPIR_Type_create_darray(int size, int rank, int ndims,
     goto fn_exit;
 }
 
-int MPIR_Type_create_subarray(int ndims, const int array_of_sizes[],
-                              const int array_of_subsizes[], const int array_of_starts[],
-                              int order, MPI_Datatype oldtype, MPI_Datatype * newtype)
+int MPIR_Type_create_subarray_impl(int ndims, const int array_of_sizes[],
+                                   const int array_of_subsizes[], const int array_of_starts[],
+                                   int order, MPI_Datatype oldtype, MPI_Datatype * newtype)
 {
     int mpi_errno = MPI_SUCCESS;
     int i;
@@ -1392,12 +1130,9 @@ void MPIR_Type_free_impl(MPI_Datatype * datatype)
     *datatype = MPI_DATATYPE_NULL;
 }
 
-int MPIR_Type_get_contents(MPI_Datatype datatype,
-                           int max_integers,
-                           int max_addresses,
-                           int max_datatypes,
-                           int array_of_integers[],
-                           MPI_Aint array_of_addresses[], MPI_Datatype array_of_datatypes[])
+int MPIR_Type_get_contents_impl(MPI_Datatype datatype, int max_integers, int max_addresses,
+                                int max_datatypes, int array_of_integers[],
+                                MPI_Aint array_of_addresses[], MPI_Datatype array_of_datatypes[])
 {
     int i, mpi_errno;
     MPIR_Datatype *dtp;
@@ -1421,7 +1156,7 @@ int MPIR_Type_get_contents(MPI_Datatype datatype,
     /* --BEGIN ERROR HANDLING-- */
     if (max_integers < cp->nr_ints || max_addresses < cp->nr_aints || max_datatypes < cp->nr_types) {
         mpi_errno = MPIR_Err_create_code(MPI_SUCCESS, MPIR_ERR_RECOVERABLE,
-                                         "MPIR_Type_get_contents", __LINE__,
+                                         "MPIR_Type_get_contents_impl", __LINE__,
                                          MPI_ERR_OTHER, "**dtype", 0);
         return mpi_errno;
     }
@@ -1449,8 +1184,7 @@ int MPIR_Type_get_contents(MPI_Datatype datatype,
     return MPI_SUCCESS;
 }
 
-void MPIR_Type_get_envelope(MPI_Datatype datatype,
-                            int *num_integers,
+void MPIR_Type_get_envelope(MPI_Datatype datatype, int *num_integers,
                             int *num_addresses, int *num_datatypes, int *combiner)
 {
     if (HANDLE_IS_BUILTIN(datatype) ||
@@ -1526,35 +1260,6 @@ void MPIR_Type_get_true_extent_x_impl(MPI_Datatype datatype, MPI_Count * true_lb
         *true_lb = datatype_ptr->true_lb;
         *true_extent = datatype_ptr->true_ub - datatype_ptr->true_lb;
     }
-}
-
-int MPIR_Type_hvector_impl(int count, int blocklength, MPI_Aint stride, MPI_Datatype oldtype,
-                           MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPI_Datatype new_handle;
-    MPIR_Datatype *new_dtp;
-    int ints[2];
-
-    mpi_errno = MPIR_Type_vector(count, blocklength, (MPI_Aint) stride, 1,      /* stride in bytes */
-                                 oldtype, &new_handle);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    ints[0] = count;
-    ints[1] = blocklength;
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_HVECTOR, 2,    /* ints (count, blocklength) */
-                                           1,   /* aints */
-                                           1,   /* types */
-                                           ints, &stride, &oldtype);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
 }
 
 int MPIR_Type_indexed(int count,
@@ -2094,47 +1799,6 @@ int MPIR_Type_struct(int count,
     goto fn_exit;
 }
 
-int MPIR_Type_struct_impl(int count, const int *array_of_blocklengths,
-                          const MPI_Aint * array_of_displacements,
-                          const MPI_Datatype * array_of_types, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPI_Datatype new_handle;
-    int i, *ints;
-    MPIR_Datatype *new_dtp;
-    MPIR_CHKLMEM_DECL(1);
-
-    mpi_errno = MPIR_Type_struct(count,
-                                 array_of_blocklengths,
-                                 array_of_displacements, array_of_types, &new_handle);
-    MPIR_ERR_CHECK(mpi_errno);
-
-
-    MPIR_CHKLMEM_MALLOC(ints, int *, (count + 1) * sizeof(int), mpi_errno, "contents integer array",
-                        MPL_MEM_BUFFER);
-
-    ints[0] = count;
-    for (i = 0; i < count; i++) {
-        ints[i + 1] = array_of_blocklengths[i];
-    }
-
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_STRUCT, count + 1,     /* ints (count, blocklengths) */
-                                           count,       /* aints (displacements) */
-                                           count,       /* types */
-                                           ints, array_of_displacements, array_of_types);
-
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-
-  fn_exit:
-    MPIR_CHKLMEM_FREEALL();
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 int MPIR_Type_vector(int count,
                      int blocklength,
                      MPI_Aint stride,
@@ -2274,6 +1938,255 @@ int MPIR_Type_vector_impl(int count, int blocklength, int stride, MPI_Datatype o
     MPIR_ERR_CHECK(mpi_errno);
 
     MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
+
+  fn_exit:
+    return mpi_errno;
+  fn_fail:
+    goto fn_exit;
+}
+
+int MPIR_Type_create_resized_impl(MPI_Datatype oldtype, MPI_Aint lb, MPI_Aint extent,
+                                  MPI_Datatype * newtype)
+{
+    int mpi_errno = MPI_SUCCESS;
+    MPI_Datatype new_handle;
+    MPIR_Datatype *new_dtp;
+    MPI_Aint aints[2];
+
+    mpi_errno = MPIR_Type_create_resized(oldtype, lb, extent, &new_handle);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    aints[0] = lb;
+    aints[1] = extent;
+
+    MPIR_Datatype_get_ptr(new_handle, new_dtp);
+    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_RESIZED, 0, 2, /* Aints */
+                                           1, NULL, aints, &oldtype);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
+
+  fn_exit:
+    return mpi_errno;
+  fn_fail:
+    goto fn_exit;
+}
+
+int MPIR_Type_dup(MPI_Datatype oldtype, MPI_Datatype * newtype)
+{
+    int mpi_errno = MPI_SUCCESS;
+    MPIR_Datatype *new_dtp = 0, *old_dtp;
+
+    if (HANDLE_IS_BUILTIN(oldtype)) {
+        /* create a new type and commit it. */
+        mpi_errno = MPIR_Type_contiguous(1, oldtype, newtype);
+        MPIR_ERR_CHECK(mpi_errno);
+    } else {
+        /* allocate new datatype object and handle */
+        new_dtp = (MPIR_Datatype *) MPIR_Handle_obj_alloc(&MPIR_Datatype_mem);
+        if (!new_dtp) {
+            /* --BEGIN ERROR HANDLING-- */
+            mpi_errno = MPIR_Err_create_code(MPI_SUCCESS, MPIR_ERR_RECOVERABLE,
+                                             "MPIR_Type_dup", __LINE__, MPI_ERR_OTHER,
+                                             "**nomem", 0);
+            goto fn_fail;
+            /* --END ERROR HANDLING-- */
+        }
+
+        MPIR_Datatype_get_ptr(oldtype, old_dtp);
+
+        /* fill in datatype */
+        MPIR_Object_set_ref(new_dtp, 1);
+        /* new_dtp->handle is filled in by MPIR_Handle_obj_alloc() */
+        new_dtp->is_contig = old_dtp->is_contig;
+        new_dtp->size = old_dtp->size;
+        new_dtp->extent = old_dtp->extent;
+        new_dtp->ub = old_dtp->ub;
+        new_dtp->lb = old_dtp->lb;
+        new_dtp->true_ub = old_dtp->true_ub;
+        new_dtp->true_lb = old_dtp->true_lb;
+        new_dtp->alignsize = old_dtp->alignsize;
+        new_dtp->is_committed = old_dtp->is_committed;
+
+        new_dtp->attributes = NULL;     /* Attributes are copied in the
+                                         * top-level MPI_Type_dup routine */
+        new_dtp->name[0] = 0;   /* The Object name is not copied on
+                                 * a dup */
+        new_dtp->n_builtin_elements = old_dtp->n_builtin_elements;
+        new_dtp->builtin_element_size = old_dtp->builtin_element_size;
+        new_dtp->basic_type = old_dtp->basic_type;
+
+        new_dtp->typerep.handle = NULL;
+        *newtype = new_dtp->handle;
+
+        if (old_dtp->is_committed) {
+            MPID_Type_commit_hook(new_dtp);
+        }
+
+        mpi_errno = MPIR_Typerep_create_dup(oldtype, new_dtp);
+        MPIR_ERR_CHECK(mpi_errno);
+    }
+
+    MPL_DBG_MSG_D(MPIR_DBG_DATATYPE, VERBOSE, "dup type %x created.", *newtype);
+
+  fn_fail:
+    return mpi_errno;
+}
+
+int MPIR_Type_dup_impl(MPI_Datatype oldtype, MPI_Datatype * newtype)
+{
+    int mpi_errno = MPI_SUCCESS;
+    MPI_Datatype new_handle;
+    MPIR_Datatype *new_dtp;
+
+    mpi_errno = MPIR_Type_dup(oldtype, &new_handle);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    MPIR_Datatype_get_ptr(new_handle, new_dtp);
+    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_DUP, 0,        /* ints */
+                                           0,   /* aints */
+                                           1,   /* types */
+                                           NULL, NULL, &oldtype);
+
+    mpi_errno = MPIR_Type_commit_impl(&new_handle);
+    MPIR_ERR_CHECK(mpi_errno);
+
+    /* Copy attributes, executing the attribute copy functions */
+    /* This accesses the attribute dup function through the perprocess
+     * structure to prevent type_dup from forcing the linking of the
+     * attribute functions.  The actual function is (by default)
+     * MPIR_Attr_dup_list
+     */
+    MPIR_Datatype *old_dtp;
+    MPIR_Datatype_get_ptr(oldtype, old_dtp);
+    if (MPIR_Process.attr_dup) {
+        new_dtp->attributes = 0;
+        mpi_errno = MPIR_Process.attr_dup(oldtype, old_dtp->attributes, &new_dtp->attributes);
+        if (mpi_errno) {
+            MPIR_Datatype_ptr_release(new_dtp);
+            goto fn_fail;
+        }
+    }
+
+    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
+
+  fn_exit:
+    return mpi_errno;
+  fn_fail:
+    goto fn_exit;
+}
+
+int MPIR_Type_match_size_impl(int typeclass, int size, MPI_Datatype * datatype)
+{
+    int mpi_errno = MPI_SUCCESS;
+#ifdef HAVE_ERROR_CHECKING
+    const char *tname;
+#endif
+    /* Note that all of the datatype have values, even if the type is
+     * not available. We test for that case separately.  We also
+     * prefer the Fortran types to the C type, if they are available */
+    static MPI_Datatype real_types[] = {
+        MPI_REAL4, MPI_REAL8, MPI_REAL16,
+        MPI_REAL, MPI_DOUBLE_PRECISION,
+        MPI_FLOAT, MPI_DOUBLE, MPI_LONG_DOUBLE
+    };
+    static MPI_Datatype int_types[] = {
+        MPI_INTEGER1, MPI_INTEGER2, MPI_INTEGER4, MPI_INTEGER8, MPI_INTEGER16,
+        MPI_INTEGER,
+        MPI_CHAR, MPI_SHORT, MPI_INT,
+        MPI_LONG, MPI_LONG_LONG
+    };
+    static MPI_Datatype complex_types[] = {
+        MPI_COMPLEX8, MPI_COMPLEX16, MPI_COMPLEX32,
+        MPI_COMPLEX, MPI_DOUBLE_COMPLEX,
+        MPI_C_COMPLEX, MPI_C_DOUBLE_COMPLEX, MPI_C_LONG_DOUBLE_COMPLEX,
+    };
+    MPI_Datatype matched_datatype = MPI_DATATYPE_NULL;
+    int i;
+    MPI_Aint tsize;
+
+    /* The following implementation follows the suggestion in the
+     * MPI-2 standard.
+     * The version in the MPI-2 spec makes use of the Fortran optional types;
+     * currently, we don't support these from C (see mpi.h.in).
+     * Thus, we look at the candidate types and make use of the first fit.
+     * Note that the standard doesn't require that this routine return
+     * any particular choice of MPI datatype; e.g., it is not required
+     * to return MPI_INTEGER4 if a 4-byte integer is requested.
+     */
+    switch (typeclass) {
+        case MPI_TYPECLASS_REAL:
+            {
+                int nRealTypes = sizeof(real_types) / sizeof(MPI_Datatype);
+#ifdef HAVE_ERROR_CHECKING
+                tname = "MPI_TYPECLASS_REAL";
+#endif
+                for (i = 0; i < nRealTypes; i++) {
+                    if (real_types[i] == MPI_DATATYPE_NULL) {
+                        continue;
+                    }
+                    MPIR_Datatype_get_size_macro(real_types[i], tsize);
+                    if (tsize == size) {
+                        matched_datatype = real_types[i];
+                        break;
+                    }
+                }
+            }
+            break;
+        case MPI_TYPECLASS_INTEGER:
+            {
+                int nIntTypes = sizeof(int_types) / sizeof(MPI_Datatype);
+#ifdef HAVE_ERROR_CHECKING
+                tname = "MPI_TYPECLASS_INTEGER";
+#endif
+                for (i = 0; i < nIntTypes; i++) {
+                    if (int_types[i] == MPI_DATATYPE_NULL) {
+                        continue;
+                    }
+                    MPIR_Datatype_get_size_macro(int_types[i], tsize);
+                    if (tsize == size) {
+                        matched_datatype = int_types[i];
+                        break;
+                    }
+                }
+            }
+            break;
+        case MPI_TYPECLASS_COMPLEX:
+            {
+                int nComplexTypes = sizeof(complex_types) / sizeof(MPI_Datatype);
+#ifdef HAVE_ERROR_CHECKING
+                tname = "MPI_TYPECLASS_COMPLEX";
+#endif
+                for (i = 0; i < nComplexTypes; i++) {
+                    if (complex_types[i] == MPI_DATATYPE_NULL) {
+                        continue;
+                    }
+                    MPIR_Datatype_get_size_macro(complex_types[i], tsize);
+                    if (tsize == size) {
+                        matched_datatype = complex_types[i];
+                        break;
+                    }
+                }
+            }
+            break;
+        default:
+            /* --BEGIN ERROR HANDLING-- */
+            MPIR_ERR_SETANDSTMT(mpi_errno, MPI_ERR_ARG, break, "**typematchnoclass");
+            /* --END ERROR HANDLING-- */
+    }
+
+    if (mpi_errno == MPI_SUCCESS) {
+        if (matched_datatype == MPI_DATATYPE_NULL) {
+            /* --BEGIN ERROR HANDLING-- */
+            MPIR_ERR_SETANDSTMT2(mpi_errno, MPI_ERR_ARG,;, "**typematchsize",
+                                 "**typematchsize %s %d", tname, size);
+            /* --END ERROR HANDLING-- */
+        } else {
+            *datatype = matched_datatype;
+        }
+    }
+    if (mpi_errno != MPI_SUCCESS)
+        goto fn_fail;
 
   fn_exit:
     return mpi_errno;
