--- src/mpi/datatype/get_elements.c
@@ -1,125 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Get_elements */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Get_elements = PMPI_Get_elements
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Get_elements  MPI_Get_elements
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Get_elements as PMPI_Get_elements
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Get_elements(const MPI_Status * status, MPI_Datatype datatype, int *count)
-    __attribute__ ((weak, alias("PMPI_Get_elements")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Get_elements
-#define MPI_Get_elements PMPI_Get_elements
-#endif
-
-/*@
-   MPI_Get_elements - Returns the number of basic elements
-                      in a datatype
-
-Input Parameters:
-+ status - return status of receive operation (Status)
-- datatype - datatype used by receive operation (handle)
-
-Output Parameters:
-. count - number of received basic elements (integer)
-
-   Notes:
-
- If the size of the datatype is zero and the amount of data returned as
- determined by 'status' is also zero, this routine will return a count of
- zero.  This is consistent with a clarification made by the MPI Forum.
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-@*/
-int MPI_Get_elements(const MPI_Status * status, MPI_Datatype datatype, int *count)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPI_Count count_x, byte_count;
-
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_GET_ELEMENTS);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPI_GET_ELEMENTS);
-
-    /* Validate parameters, especially handles needing to be converted */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            MPIR_ERRTEST_DATATYPE(datatype, "datatype", mpi_errno);
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif
-
-    /* Validate parameters and objects (post conversion) */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            MPIR_Datatype *datatype_ptr = NULL;
-
-            MPIR_ERRTEST_ARGNULL(status, "status", mpi_errno);
-            MPIR_ERRTEST_ARGNULL(count, "count", mpi_errno);
-            /* Convert MPI object handles to object pointers */
-            MPIR_Datatype_get_ptr(datatype, datatype_ptr);
-            /* Validate datatype_ptr */
-            if (!HANDLE_IS_BUILTIN(datatype)) {
-                MPIR_Datatype_get_ptr(datatype, datatype_ptr);
-                MPIR_Datatype_valid_ptr(datatype_ptr, mpi_errno);
-                if (mpi_errno != MPI_SUCCESS)
-                    goto fn_fail;
-
-                MPIR_Datatype_committed_ptr(datatype_ptr, mpi_errno);
-                if (mpi_errno != MPI_SUCCESS)
-                    goto fn_fail;
-            }
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif
-
-    /* ... body of routine ...  */
-
-    byte_count = MPIR_STATUS_GET_COUNT(*status);
-    mpi_errno = MPIR_Get_elements_x_impl(&byte_count, datatype, &count_x);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* clip the value if it cannot be correctly returned to the user */
-    *count = (count_x > INT_MAX) ? MPI_UNDEFINED : (int) count_x;
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPI_GET_ELEMENTS);
-    return mpi_errno;
-
-    /* --BEGIN ERROR HANDLING-- */
-  fn_fail:
-    {
-        mpi_errno =
-            MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
-                                 "**mpi_get_elements", "**mpi_get_elements %p %D %p", status,
-                                 datatype, count);
-    }
-    mpi_errno = MPIR_Err_return_comm(0, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
