--- src/mpi/datatype/type_create_hvector.c
@@ -1,128 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Type_create_hvector */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Type_create_hvector = PMPI_Type_create_hvector
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Type_create_hvector  MPI_Type_create_hvector
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Type_create_hvector as PMPI_Type_create_hvector
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Type_create_hvector(int count, int blocklength, MPI_Aint stride, MPI_Datatype oldtype,
-                            MPI_Datatype * newtype)
-    __attribute__ ((weak, alias("PMPI_Type_create_hvector")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Type_create_hvector
-#define MPI_Type_create_hvector PMPI_Type_create_hvector
-
-#endif
-
-/*@
-   MPI_Type_create_hvector - Create a datatype with a constant stride given
-     in bytes
-
-Input Parameters:
-+ count - number of blocks (nonnegative integer)
-. blocklength - number of elements in each block (nonnegative integer)
-. stride - number of bytes between start of each block (address integer)
-- oldtype - old datatype (handle)
-
-Output Parameters:
-. newtype - new datatype (handle)
-
-.N ThreadSafe
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_TYPE
-.N MPI_ERR_ARG
-@*/
-int MPI_Type_create_hvector(int count,
-                            int blocklength,
-                            MPI_Aint stride, MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPI_Datatype new_handle;
-    MPIR_Datatype *new_dtp;
-    int ints[2];
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TYPE_CREATE_HVECTOR);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPI_TYPE_CREATE_HVECTOR);
-
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            MPIR_Datatype *datatype_ptr = NULL;
-
-            MPIR_ERRTEST_COUNT(count, mpi_errno);
-            MPIR_ERRTEST_ARGNEG(blocklength, "blocklen", mpi_errno);
-            MPIR_ERRTEST_DATATYPE(oldtype, "datatype", mpi_errno);
-
-            if (!HANDLE_IS_BUILTIN(oldtype)) {
-                MPIR_Datatype_get_ptr(oldtype, datatype_ptr);
-                MPIR_Datatype_valid_ptr(datatype_ptr, mpi_errno);
-                if (mpi_errno != MPI_SUCCESS)
-                    goto fn_fail;
-            }
-            MPIR_ERRTEST_ARGNULL(newtype, "newtype", mpi_errno);
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ... */
-
-    mpi_errno = MPIR_Type_vector(count, blocklength, stride, 1, /* stride in bytes */
-                                 oldtype, &new_handle);
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    ints[0] = count;
-    ints[1] = blocklength;
-    MPIR_Datatype_get_ptr(new_handle, new_dtp);
-    mpi_errno = MPIR_Datatype_set_contents(new_dtp, MPI_COMBINER_HVECTOR, 2,    /* ints (count, blocklength) */
-                                           1,   /* aints */
-                                           1,   /* types */
-                                           ints, &stride, &oldtype);
-
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    MPIR_OBJ_PUBLISH_HANDLE(*newtype, new_handle);
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPI_TYPE_CREATE_HVECTOR);
-    MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        mpi_errno = MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__,
-                                         MPI_ERR_OTHER, "**mpi_type_create_hvector",
-                                         "**mpi_type_create_hvector %d %d %L %D %p", count,
-                                         blocklength, stride, oldtype, newtype);
-    }
-#endif
-    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
