--- src/mpi/datatype/type_vector.c
@@ -1,111 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-#include "datatype.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Type_vector */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Type_vector = PMPI_Type_vector
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Type_vector  MPI_Type_vector
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Type_vector as PMPI_Type_vector
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Type_vector(int count, int blocklength, int stride, MPI_Datatype oldtype,
-                    MPI_Datatype * newtype) __attribute__ ((weak, alias("PMPI_Type_vector")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Type_vector
-#define MPI_Type_vector PMPI_Type_vector
-#endif
-
-/*@
-    MPI_Type_vector - Creates a vector (strided) datatype
-
-Input Parameters:
-+ count - number of blocks (nonnegative integer)
-. blocklength - number of elements in each block
-  (nonnegative integer)
-. stride - number of elements between start of each block (integer)
-- oldtype - old datatype (handle)
-
-Output Parameters:
-. newtype - new datatype (handle)
-
-.N ThreadSafe
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_ARG
-.N MPI_ERR_OTHER
-@*/
-int MPI_Type_vector(int count,
-                    int blocklength, int stride, MPI_Datatype oldtype, MPI_Datatype * newtype)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TYPE_VECTOR);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPI_TYPE_VECTOR);
-
-    /* Validate parameters, especially handles needing to be converted */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            MPIR_Datatype *old_ptr = NULL;
-
-            MPIR_ERRTEST_COUNT(count, mpi_errno);
-            MPIR_ERRTEST_ARGNEG(blocklength, "blocklen", mpi_errno);
-            MPIR_ERRTEST_DATATYPE(oldtype, "datatype", mpi_errno);
-
-            if (oldtype != MPI_DATATYPE_NULL && !HANDLE_IS_BUILTIN(oldtype)) {
-                MPIR_Datatype_get_ptr(oldtype, old_ptr);
-                MPIR_Datatype_valid_ptr(old_ptr, mpi_errno);
-                if (mpi_errno)
-                    goto fn_fail;
-            }
-            MPIR_ERRTEST_ARGNULL(newtype, "newtype", mpi_errno);
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ...  */
-
-    mpi_errno = MPIR_Type_vector_impl(count, blocklength, stride, oldtype, newtype);
-    if (mpi_errno)
-        goto fn_fail;
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPI_TYPE_VECTOR);
-    MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        mpi_errno =
-            MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
-                                 "**mpi_type_vector", "**mpi_type_vector %d %d %d %D %p", count,
-                                 blocklength, stride, oldtype, newtype);
-    }
-#endif
-    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
